+++ src/main/java/com/example/concertreservationsystem/application/concert/facade/ConcertFacadeImpl.java
@@ -2,6 +2,7 @@ package com.example.concertreservationsystem.application.concert.facade;
 
 import com.example.concertreservationsystem.application.concert.dto.request.ConcertRequestDto;
 import com.example.concertreservationsystem.application.concert.dto.response.ConcertResponseDto;
+import com.example.concertreservationsystem.domain.service.concert.ConcertService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
