+++ src/main/java/com/example/concertreservationsystem/application/event/facade/ConcertEventFacadeImpl.java
@@ -2,6 +2,7 @@ package com.example.concertreservationsystem.application.event.facade;
 
 import com.example.concertreservationsystem.application.event.dto.request.EventRequestDto;
 import com.example.concertreservationsystem.application.event.dto.response.EventResponseDto;
+import com.example.concertreservationsystem.domain.service.event.ConcertEventService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
