+++ src/main/java/com/example/concertreservationsystem/application/queue/facade/QueueFacadeImpl.java
@@ -1,6 +1,7 @@
 package com.example.concertreservationsystem.application.queue.facade;
 
 import com.example.concertreservationsystem.application.queue.dto.QueueResponseToken;
+import com.example.concertreservationsystem.domain.service.queue.QueueService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
