--- src/main/java/com/example/concertreservationsystem/application/usecase/ConcertUseCase.java
@@ -1,13 +1,0 @@
-package com.example.concertreservationsystem.application.usecase;
-
-
-import com.example.concertreservationsystem.application.concert.dto.request.ConcertRequestDto;
-import com.example.concertreservationsystem.application.concert.dto.response.ConcertResponseDto;
-import org.springframework.stereotype.Component;
-
-@Component
-public interface ConcertUseCase {
-    public ConcertResponseDto registerConcert(ConcertRequestDto requestDto);
-
-}
-
