--- src/main/java/com/example/concertreservationsystem/application/usecase/QueueUseCase.java
@@ -1,11 +1,0 @@
-package com.example.concertreservationsystem.application.usecase;
-
-
-import com.example.concertreservationsystem.application.queue.dto.QueueResponseToken;
-import org.springframework.stereotype.Component;
-
-@Component
-public interface QueueUseCase {
-
-    public QueueResponseToken addQueueToUser(String uuid);
-}
