+++ src/test/java/com/example/concertreservationsystem/concert/ConcertUnitTest.java
@@ -2,7 +2,7 @@ package com.example.concertreservationsystem.concert;
 
 import com.example.concertreservationsystem.domain.model.Concert;
 import com.example.concertreservationsystem.domain.repo.ConcertRepository;
-import com.example.concertreservationsystem.application.concert.facade.ConcertService;
+import com.example.concertreservationsystem.domain.service.concert.ConcertService;
 import com.example.concertreservationsystem.application.concert.dto.request.ConcertRequestDto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
